added database::init

+++ migrations/20240408162944_init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
-    id BIGINT PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
     username TEXT NOT NULL UNIQUE,
     password TEXT NOT NULL
 );
@@ -9,7 +9,7 @@ VALUES (
         '89ba60446ddfb9f296863aaa0d6431305fa87c78058d674466672f780be9ecef'
     ) ON CONFLICT (username) DO NOTHING;
 CREATE TABLE IF NOT EXISTS market_data (
-    id BIGINT PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
     exchange TEXT NOT NULL,
     symbol TEXT NOT NULL,
     market_data_type TEXT NOT NULL,
